chore: revert changes introduced by allowing stopping services in starting state (#510)
In [this PR](https://github.com/canonical/pebble/pull/503), we
introduced a feature to allow stopping services that are in the starting
state / within the okayDelay. However, this makes [a deadlock issue in
ringbuffer a real
issue](https://github.com/canonical/pebble/issues/508). So we are
reverting this change now and will redo it after the deadlock is
resolved.