feat(run): Add the --dry option to the run command
The --dry option changes the behavior of the run command to only execute
the setting-up of the pebble daemon without actually starting it.
Side-effects of initialization on the state file are also disabled.

This allows checking the validity of a plan without launching pebble
proper.
For example, if you have a plan layer at /tmp/layers/123-layer.yaml, you
can check it using `PEBBLE=/tmp go run ./cmd/pebble run --dry` provided
this command is invoked from the root of this repository.