feat(workloads): add plumbing to support workloads
This is a preliminar patch that adds support for workloads in plan,
plan manager and service manager by passing a workload.Provider instance
from the daemon.New() constructor all the way down to servstate manager.
This also requires splitting the plan.ReadDir() and (*Plan).Validate()
methods, and passing the workload.Provider as an argument to Validate(),
though this is a temporary measure to get the workloads feature working
before refactoring the plan and service manager code to make services
a plan section, rather than an ad-hoc member of the Plan struct.

In the future, this patch needs to be refactored to store the Provider
instance in a future service extension, when the services layer in the
plan is refactored to be a plan section. This way, we can pass the
workload.Provider from a plan extension and keep the current plan,
daemon, and overlord APIs intact.