docs: change Diataxis categories to the standard order (#559)
The Diataxis categories in Pebble are currently ordered: Tutorial,
How-to guides, Explanation, Reference. The standard order should be
Tutorial, How-to guides, _Reference, Explanation_.

The 2x2 grid of cards on the homepage actually has the correct layout
(matching the [Diataxis map](https://diataxis.fr/_images/diataxis.png)):

```
Tutorial     How-to guides
Explanation  Reference
```

This PR:

- Reorders the Diataxis categories to Tutorial, How-to guides,
Reference, Explanation.
- Uses `:reverse:` in the grid source to swap Reference and Explanation,
instead of ordering them as Explanation > Reference in the grid source.
We need to do it this way so that Reference stacks above Explanation if
the browser window is narrow.

In short, we're matching how Ops is set up. See
https://github.com/canonical/operator/blob/main/docs/index.md?plain=1#L54