chore: bump the upload-artifact and download-artifact version (#566)
[v3 of upload-artifact and download-artifact can no longer be
used](https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/).
This PR bumps the version to v4.

For upload, v4 differs from v3 in that v3 would allow uploading multiple
times with the same name and v4 does not. We do re-use the
snapcraft-logs name - but only in failure, so I *think* this means that
the workflow would exit before any clash occurred. We could adjust the
names to avoid that possibility - I'm guessing that we only manually
inspect them, rather than use the result in future work.