feat: add a startup field for checks, and start-checks and stop-checks actions (#560)
Add support for starting and stopping checks, in a very similar manner
to starting and stopping services. This includes a new optional layer
field for checks, `startup`, that determines whether or not the check
will be started when added to the plan.

I've manually tested this outside of the edge cases (for example, I've
added a plan with the three `startup` values and checked they behave as
expected both then and in a replan, I've started and stopped those
checks both when all in the 'wrong' state, all already in the desired
state, and a mixture of those, I've tried to start/stop checks that
don't exist, I've changed the period for started and stopped checks and
made sure the behaviour is correct, and so forth). I've done the
majority of this with a locally running Pebble and the CLI, and the core
checks also with a locally running Pebble and the Python Pebble client
with the draft support for this.

### Layer

The layer specification gains a new `startup` field for checks, which
must be either `enabled` or `disabled`, and defaults to `enabled` for
backwards compatibility.

The impact of this field is that when the plan is updated (including
when Pebble first starts) or a layer is added, new checks where
`startup` is set to `disabled` will not be started. If a check's
configuration changes (for example, the timeout or period), and it is
set to `startup: disabled` but is currently running, then it will be
restarted as previously. Any stopped checks where `startup` is changed
to `enabled` will also be started when the plan changes.

This field is implemented via the existing check manager `PlanChanged`
function. That function is registered as a listener for when the plan
changes, which includes new checks being added, and check fields
(including `startup`) changing. The function is adjusted in this PR to
only start new checks if `startup` is enabled.

### CLI

Minor change: since `replan` now impacts checks as well as services, it
is moved in the help listing from the `services` section to the `plan`
section.

Two new commands are added: `start-checks` and `stop-checks`. These take
one or more check names as arguments, and start or stop the specified
checks. Starting a running check or stopping a stopped check is a no-op.
The command response informs the user which checks actually required
stopping or starting.

The `checks` command output is altered to include a new "Startup"
column, which is simply the value of the field from the combined plan
for that check, and to show `inactive` in the "Status" column for any
checks that have been stopped (these also get a `-` value for both
"Change" and "Failures").

### API

The existing `/v1/checks` endpoint is extended to also support `POST`,
requiring admin access. The structure for this endpoint is similar to
posting to `/v1/services`, in that it requires a list of check names and
an action to perform (currently either `start` or `stop`). The response
is a `SyncResponse` that contains a payload that has a `changed` list of
checks that did require starting/stopping. The API waits for the checks
to start/stop before returning.

In addition, the `replan` command of the `/v1/services` endpoint now
calls the `Replan` function in the check manager, which effectively does
an "autostart" for checks.

[Spec -
OP052](https://docs.google.com/document/d/1KpgFR0eaxkrwt_Vd0rTZzb0D57dQVtAv7t1b9iuMUQw/edit?tab=t.0)
(internal link only, sorry).

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>