feat: sort the changed checks in the API response (#572)
The PR changes the response to a `/v1/checks` POST to provide the list
of names of started or stopped checks as sorted alphabetically.

A test is added to confirm this, which also tests the API post method
more generally, which was missed in #560

This will also provide a sorted list of check names in the client (since
it uses the API) and the CLI (since it uses the client). The *manager*
continues to return the list of check names in the order in which they
were started/stopped - this feels more correct to me (in terms of
management, what matters is what happened; in terms of response to a
request, convenience is important), but I could be convinced that the
sorting should move up a level.