fix: avoid health checks URL panic; add staticcheck (#580)
This fixes #579, which results in a panic if an HTTP check URL was
malformed. Thanks for the report and the test at
https://github.com/canonical/tako/pull/24, @alnvdl-work.

I also fixed another similar "forgot to check err" issue in
`cmd_warnings.go`.

This PR also adds [staticcheck](https://staticcheck.dev/) to the
linting, which found this issue, and a few other minor issues. Fix the
other issues and add a `staticcheck.conf` and a GitHub Action to run it.
The other issues are mostly not problems, but include:

- Use of deprecated functions (these are okay, just `lint:ignore`)
- Error messages shouldn't end with "."
- `strconv.ParseFloat` second arg is not the base (10), so should be 32
or 64. I actually
[tried](https://go-review.googlesource.com/c/tools/+/267600) to add this
check to `go vet` years ago.
- Various unused functions and struct fields.
- `time.Now().Sub(t)` -> `time.Since(t)`
- Use of `rand.Seed` (deprecated) in tests

Fixes #579