docs: improve description for Snap Store listing (#589)
This PR improves the description that appears on the [Snap Store
page](https://snapcraft.io/pebble) for Pebble.

Note that I removed mention of `SNAP_USER_DATA` by name, as that
variable is not set outside of the snap. Outside of the snap, the most
relevant point is that layer config is stored in `$PEBBLE/layers`, where
`PEBBLE` is set to `$HOME/snap/pebble/<snap_revision>` by default.