feat(idkey): add identity key [1/4] (#605)
Add support for Pebble to have a machine, container or device identity
that is based on an Ed25519 private key.

This identity key implements the crypto.Signer interface, allowing it to
sign, for example, TLS certificates.

The identity key should eventually be stored in a hardware backed
key-store, such as a TPM, OP-TEE or even YubiKey. The first
implementation of the package implements disk based storage with
restricted permissions, similar to what you would find when looking at
your ~/.ssh directory.

The purpose of this package is to allow a machine, container or device
to have a lifelong identity, including a standardized fingerprint that
represents that identity.

**Example use-case**

The identity private key can be used to sign short-lived TLS keypairs
for the HTTPS Pebble API server, whereby it acts as the signing
authority (CA). In this case, an x509 identity certificate with CA
signing permissions will be included in the certificate chain, allowing
the client to pin the identity certificate (which does not change over
time) during a physical trust exchange process (think button press and
user assisted identity confirmation).

**Benchmark information**

The identity will only be generated on first boot, and will incur the
following startup penalty:

 **3.0 ms** - AMD FX(tm)-8350 Eight-Core Processor
**5.5ms** - Raspberry Pi 5 (Sdcard)
**14.5ms** - Raspberry Pi 3b+ (Sdcard)