feat(tls): add TLS keypair manager [2/4] (#596)
**Add a TLS keypair manager**

The TLS manager relies on an externally supplied identity key for
signing (crypto.Signer). The implementation can support hardware backed
keystores, such as TPM, OP-TEE or even Yubikey. The current ```idkey```
package supplies a key implementation which for now persists the key on
disk.

The TLS manager will generate and persist an identity certificate that
acts as the root CA. The identity certificate is always included in the
server supplied certificate chain to the client. The actual Pebble
use-case must design a trust exchange process whereby the client can
obtain and trust (pin) the identity certificate. The client can then use
the identity certificate to validate all future TLS certificates from
the same server, while the server holds the same identity key
(crypto.Signer).

The TLS manager keeps one active TLS keypair in memory. A TLS keypair is
only ever generated on demand, and can be reused while valid. On expiry
of the TLS keypair, the manager will replace the in-memory TLS keypair
with a new one. TLS keypair rotation is seamlessly handled by the TLS
manager.

**Benchmarking**

The worse case on-demand delay for issuing the certificate chain to the
client during the TLS handshake happens when both the identity
certificate has to be created the first time, and a new TLS keypair must
be created and signed:

**4.1ms**- AMD FX(tm)-8350 Eight-Core Processor
**6.0ms** - Raspberry Pi 5 (Sdcard)
**20ms** - Raspberry Pi 3b+ (Sdcard)

Note that once the TLS certificate is valid, no additional TLS overhead
exists during future session handshakes.

**Example Usage**

1. Start the TLS manager, supplying it with a crypto.Signer (see
```idkey``` package) for the identity key.

```go
    mgr, err := tlsstate.NewManager(tlsDir, key)
    if err != nil {
        return err
    }
```

2. On demand in-memory TLS keypairs signed with the identity for
incoming HTTPS sessions (seamless rotation)

```go
    tlsCert, err := mgr.GetCertificate(...)
    if err != nil {
        return err
    }
    // Use tlsCert to start a new incoming HTTPS session ...
```
**Next steps ...**

Work is planned to allow the TLS implementation to support mTLS. The
primary goal with mTLS is to allow the client provided certificate to
serve as the client identity, replacing the current Basic Auth
credentials.