chore: respond to review feedback
Simplify the "what notices to prune" logic, to use slices.SortFunc rather than
a heap. It changes it from O(N) to O(N log N), but the N is likely to be in the
low thousands as we prune every 10 minutes.

Clarify some naming of structures to simplify reading the code.