fix: when pruning changes preserve referential integrity for notices (#613)
When pruning changes, also prune notices under a few rules:

1) Remove Changes down to 1000 (changed from 500) according to the
existing logic.
2) Remove "change-update" Notices that refer to changes that do not
exist.
3) Further remove Notices down to 1000 Notices.
4) Log an aggregate result of what changes and notices have been pruned.

This is an update to https://github.com/canonical/pebble/pull/611, built
off of 1.19.0. Some of the key differences to 611 are:
a) We prune all notices that refer to an unknown change, not just the
ones that were recently pruned. This is better for consistency, and it
means we don't need to track an extra set of what has changed.
b) Notices are also kept to an absolute maximum count.

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>