test: fix tests when run on timezones west of UTC (#616)
The test failures happen because we're displaying local time and
matching on the same day as the UTC time, which is correct east of UTC
but not west of UTC. To repro, run:

```
TZ="America/Sao_Paulo" go test ./... -count=1
```

This pulls in the fix for `timeutil.TestScheduleNext` from
https://github.com/canonical/snapd/pull/12859. I don't *love* patching
the package variable `time.Local`, but it seems better to use the same
change as snapd (kind of our upstream).

This also fixes a similar issue in the `cli` tests. There we can't
easily patch `time.Local`, because that leads to a data race with the
testing http.Server also accessing `time.Local`. So fix these in a more
direct and concurrency-safe way.

Fixes #592.