feat: make checks report "unknown" before they've been run (#620)
Currently `/v1/checks` returns a status of "up" before a check has been
run the first time. This is misleading, as it's not really up, but
"unknown". This changes the API to return "unknown" for this state.

However, the `/v1/health` endpoint still returns healthy (true) for this
period, to avoid too a breaking change with health probes (plus, it's
not clear that that's incorrect -- a boolean is all we have there, and
it's not "down" either).

Fixes #164