docs: add sitemap.xml (#632)
This PR updates Pebble to the latest version of the [Canonical Sphinx
Starter Pack](https://github.com/canonical/sphinx-docs-starter-pack) and
enables a sitemap for the doc site.

In the preview, the sitemap is generated at:

https://canonical-ubuntu-documentation-library--632.com.readthedocs.build/pebble/sitemap.xml

After we merge this PR, the sitemap will be generated at:
https://documentation.ubuntu.com/pebble/sitemap.xml

The sitemap should be helpful for SEO. In the future, the top-level
sitemap at documentation.ubuntu.com should point down to Pebble's
sitemap (and the sitemaps for all the other doc sites under
documentation.ubuntu.com).