fix bug with successes==0 after it starts succeeding again
Also add a test assertion for this case (which fails before the fix).

This main thing here was using "details.Successes, details.Failure"
instead of "0, 0" in the call to updateCheckData in changeStatusChanged.

However, also call updateCheckData on the check immediately after it
succeeds in doRecoverCheck.

Also always call updateCheckData in doPerformCheck error case. This
glitch likely isn't observable unless, but might as well fix/simplify
it anyway.