feat: add "successes" field to checks API and CLI output (#633)
This adds a "Successes" column to the output of `pebble checks` (and
field to the `pebble check` YAML), as well as a "successes" field in the
`/v1/checks` API response.

"Successes" is the number of times this check has succeeded. It is reset
when the check succeeds again after the check's failure threshold was
reached. It will be zero if the check has never run, or has never run
successfully. If successes==0 and failures==0, the caller knows the
check has not yet been run.

The client field `CheckInfo.Successes` is typed as a pointer (`*int`) as
it's being added in a much later version of Pebble -- this means clients
will be able to tell the difference between `*Successes==0` and
`Successes==nil` (nil meaning the field is not present in the API
response because they're querying on older version of Pebble). In the
`pebble checks` CLI output the column will show `?` if running against
an older Pebble daemon. I figure we can do something similar on the
Python side (`successes: int | None`).

Example `pebble checks` output:

```
$ pebble checks
Check   Level  Startup   Status    Successes  Failures  Change
up      alive  enabled   up        5          0/1       10
```

Note that this PR is an alternative to the new status "unknown" reverted
in https://github.com/canonical/pebble/pull/628.