feat: prune service logs and serviceData in service manager (#653)
Prune service logs and `serviceData` in the service manager.

- If a service is inactive, and its `currentSince` is older than
`pruneWait`, it will be pruned.
- If the total number of entries of serviceData is still greater than
`maxServiceData`, inactive services will be pruned even if they are
_not_ older than `pruneWait` to keep the total number of serviceData
entries under control. In this case, it's sorted by `currentSince` to
ensure pruning older inactive services first.

Also added a few unit test cases to cover this feature. Test cases for
the overlord are not updated because the outer prune loop is already
tested; no integration test is added because the unit tests already
cover all cases.

Fixes #539.

---------

Co-authored-by: Ben Hoyt <benhoyt@gmail.com>