fix(docs): offer only latest Loki link (#662)
The Loki how-to instructs the user to download the sample local
configuration from the repository ```main``` while pointing to the Loki
releases page for downloading the binary. The issue is that the latest
(matching the configuration) release is not always the top entry on the
releases page, as multiple parallel versions are maintained.

The current state of using the configuration file with the top release
entry is:
```
flotter@flotter-pc:~/pebble/loki$ ./loki-linux-amd64 -config.file=loki-local-config.yaml
failed parsing config: loki-local-config.yaml: yaml: unmarshal errors:
  line 30: field enable_multi_variant_queries not found in type validation.plain. Use `-config.expand-env=true` flag if you want to expand environment variables in your config file
```

Update the link to only point to the latest (highest versioned) release,
which has a much better chance of matching the downloaded configuration
file.