chore: update TIOBE dependencies to support Go 1.24 (#669)
Our TIOBE contact let us know that we need to update the staticcheck
tool that the workflow uses, to support Go 1.24.

In order to have this work in CI, also bump the version of Go used in
the unit and integration tests to 1.24, by pulling the version from
`go.mod`. This removes the Go version matrix testing in the tests, but
we only had a single version being tested, so the matrix wasn't really
used.