fix(cmdstate): concurrent access to executions map (#686)
When using conditional variables in Go (sync.Cond), broadcasts are not
queued. If a broadcast is performed outside of a Wait(), it has no
effect on that particular consumer.

For the multiple possible consumers of ```m.executions``` to always
observe every change, a broadcast must take place either:

1. Before the ```m.executions``` Wait() loop is entered, so that the
loop entry will inspect the state.
2. During the conditional Wait(), so that an inspection of
```m.executions``` is guaranteed on the following loop iteration.

The only way to ensure this is to only ever modify ```m.executions```
**and** broadcast while having the lock.

The same logic applies to the ```stopWait``` channel closure and
broadcast.

**Note:**

The conditional is created like this:

```executionsCond: sync.NewCond(&sync.Mutex{})```

Which means we have no external reference to the mutex that Wait() unlocks. For this reason the code simply uses the methods provided by the conditional to lock and unlock. Alternatively, we can define a standalone mutex externally, but in that case we need to initialize the conditional variable with it so Wait() unlocks the correct mutex.