test(testutil): remove test using reaper
Remove redundant test case TestFakeCommandWithReaper. It appears there
is some history with properly awaiting faked commands with the old
shellcheck logic and using the reaper (see #247 and #281) but it appears
this code has been removed. Looking at the rationale for removing
shellcheck in #381, and not wanting FakeCommand to be reaper aware, this
test likely should have been removed at that time. Additionally, this
code does not appear to test anything unique compared to the
exec.Command.Run() style test as written.