test(testutil): remove test using reaper (#692)
Remove redundant test case TestFakeCommandWithReaper. 

---

It appears there is some history with properly awaiting faked commands
with the old shellcheck logic and using the reaper (see
https://github.com/canonical/pebble/pull/247 and
https://github.com/canonical/pebble/pull/281) but it appears this code
has been removed. Looking at the rationale for removing shellcheck in
https://github.com/canonical/pebble/pull/381, and not wanting
FakeCommand to be reaper aware, this test likely should have been
removed at that time. Additionally, this code does not appear to test
anything unique compared to the exec.Command.Run() style test as
written.

Theoretically this test isn't hurting anything but the meta context here
is that we would like to copy the testutil package over to
[x-go](https://github.com/canonical/x-go/) and not have extra
dependencies unless we really need it.