fix: websocket connection timeout (#691)
Fix `CommandManager.Connect` so that if the Pebble side has already
timed out (an exec change fails before the execution object is created
in the `CommandManager`), the server returns an error, causing the
connection to the websocket URL to be timed out immediately.

Also, add a unit test to cover this scenario.

Closes https://github.com/canonical/pebble/issues/422.