chore: fix failed unit test (#697)
Fix the newly added unit test `TestExecChangeReady`.

Since we merged the other PR that refactors the logging, the regex needs
to be updated. Also, revert a change in `api_tasks.go`; it is supposed
to be an internal error instead of not found. In the refactor PR, we
should only change the format of the error, not the type.