fix(plan): empty sections should support defaults (#699)
The plan library has an important property that should always be
maintained - the plan sections (built-in and registered) should always
look consistent to the plan consumers getting updates. What this means
is that a section should have a well defined zero state, which may
include default values.

This PR does not affect the built-in section types, but adds support for
post-merge default values for registered section extensions.

Let's take a hypothetical example:

```yaml
pairing:
    mode: disabled
```

In the example above, there is a default rule applied **post-merge**. If
no explicit pairing mode was supplied by any of the current plan layers,
the default is "disabled". In order to support defaults, we **always**
have to include the extensions combine call, even in the case where the
plan is empty (initially). The combine is the logical place where the
combined state can be evaluated, and some decisions can be made to apply
defaulting rules.

Note: Since ```NewPlan``` is called during Pebble startup, and because
Plan section registration happens before this, any errors happening
inside this function can be considered a coding bug, which is why I am
sticking with ```panics```.