fix: stop "exited" services to avoid spurious restarts
If a service is configured with "on-failure: shutdown" and it fails,
Pebble will shut down, but then when Pebble runs again it'll try to
redo that "start" task and immediately shut down again. The task is in
status 3 (DoingStatus).

This change fixes that bug: it simply adds stateExited to the list of
service states we stop when shutting down the daemon, and this cleans
up those start tasks in DoingStatus (they'll now be ErrorStatus).