fix: don't run "start" tasks again, to avoid spurious restarts (#709)
If a service is configured with "on-failure: shutdown" and it fails,
Pebble will shut down, but then when Pebble runs again it'll try to redo
that "start" task and immediately shut down again. The task is in status
3 (DoingStatus).

This change fixes that bug: it'll skip running "start" tasks again, if
they've been started once before (from a previous Pebble run, for
example).

This PR also adds an integration test (regression test) for that.

Fixes #527.