feat(client): mTLS support [2b/2] (#715)
The client side mTLS code was split from #708.

For the complete description of the work, please refer to the
description in #708.

Pebble client changes adds:
- The ability to supply the client identity certificate, but this is not
yet used by the CLI and therefore Pebble cannot use HTTPS today.
- The ability to obtain the server identity certificate by making an
insecure HTTPS connection. Can be used to implement an SSH-like workflow
to implement a CLI warning asking the user to trust the server identity
for future connections.
- The ability to pair with the server by supplying the server identity
fingerprint. This workflow requires obtaining the fingerprint by other
means, such as DNSSD or via a display on the server side.