chore: update doc Makefile and reformat terminal directives in doc source (#717)
This PR brings in functionality & fixes from an upcoming version of the
doc starter pack. This is a partial upgrade of the starter pack, only
bringing in changes related to building the docs. I'll do a full upgrade
at a later date.

The format of the `{terminal}` directive is changing. Instead of using
`:input:` to denote the terminal command entered, a blank line will
separate the terminal command from the output. [More detail about the
change](https://github.com/canonical/sphinx-terminal/releases/tag/1.0.0)