chore: update Rockcraft documentation links (#733)
It looks like there was some reorganisation of the Rockcraft docs, which
broke a couple of our links. The PR updates them to essentially the same
content.

The PR also switches to inter-sphinx links for Rockcraft, since most of
the links are changing in this PR anyway.

[Preview
link](https://canonical-ubuntu-documentation-library--733.com.readthedocs.build/pebble/how-to/capture-logs-from-files/).

---------

Co-authored-by: Tony Meyer <career@aotearoa.dev>