client: improve error message when pulling non-existent files (#754)
Fixed error message when attempting to pull a file that doesn't exist on
the remote system.

Previously, when a remote file didn't exist, the Pull function would
return: as mentioned in #722
```
error: expected first field name to be "files", got "response"
```

This was confusing because the API returns a "response" part (without a
"files" part) when the file doesn't exist, but the client expected
"files" first.

Now the client checks if the first multipart part is "response" and
extracts the actual error from the API, resulting in a clearer error
message:
```
error: stat /tmp/file-does-not-exist.txt: no such file or directory
```

This fix also refactors the response processing logic into a shared
`processResponsePart()` helper function to eliminate code duplication
between the error-first and success paths.

Fixes #722