ci: use named artefacts per architecture (#768)
Since we've merged download-artefact version bump and I tried to fix the
local build, the remote build began to fail, because when multiple
artefacts are available, download would download all of them, but then
unpack into a directory.

Set exact names on the artefacts, so that the test and release step
download a single snap, for the specific architecture.